+++ apis/namespaced/v1alpha1/doc.go
@@ -1,0 +1,9 @@
+/*
+Copyright 2021 Upbound Inc.
+*/
+
+// Package v1alpha1 contains the core resources of the openstack jet provider.
+// +kubebuilder:object:generate=true
+// +groupName=openstack.m.crossplane.io
+// +versionName=v1alpha1
+package v1alpha1