+++ apis/namespaced/v1beta1/doc.go
@@ -1,0 +1,9 @@
+/*
+Copyright 2022 Upbound Inc.
+*/
+
+// Package v1beta1 contains the core resources of the openstack upjet provider.
+// +kubebuilder:object:generate=true
+// +groupName=openstack.m.crossplane.io
+// +versionName=v1beta1
+package v1beta1