+++ cluster/test/setup.sh
@@ -3,13 +3,13 @@ set -aeuo pipefail
 
 echo "Running setup.sh"
 echo "Creating cloud credential secret..."
-${KUBECTL} -n upbound-system create secret generic provider-secret --from-literal=credentials="${UPTEST_CLOUD_CREDENTIALS}" --dry-run=client -o yaml | ${KUBECTL} apply -f -
+${KUBECTL} -n crossplane-system create secret generic provider-secret --from-literal=credentials="${UPTEST_CLOUD_CREDENTIALS}" --dry-run=client -o yaml | ${KUBECTL} apply -f -
 
 echo "Waiting until provider is healthy..."
 ${KUBECTL} wait provider.pkg --all --for condition=Healthy --timeout 5m
 
 echo "Waiting for all pods to come online..."
-${KUBECTL} -n upbound-system wait --for=condition=Available deployment --all --timeout=5m
+${KUBECTL} -n crossplane-system wait --for=condition=Available deployment --all --timeout=5m
 
 echo "Creating a default provider config..."
 cat <<EOF | ${KUBECTL} apply -f -
@@ -22,5 +22,21 @@ spec:
     source: Secret
     secretRef:
       name: provider-secret
-      namespace: upbound-system
+      namespace: crossplane-system
       key: credentials
+EOF
+
+echo "Creating a default provider config..."
+cat <<EOF | ${KUBECTL} apply -f -
+apiVersion: openstack.m.crossplane.io/v1beta1
+kind: ClusterProviderConfig
+metadata:
+  name: default
+spec:
+  credentials:
+    source: Secret
+    secretRef:
+      name: provider-secret
+      namespace: crossplane-system
+      key: credentials
+EOF