+++ config/cluster/lb/config.go
@@ -1,0 +1,12 @@
+package lb
+
+import "github.com/crossplane/upjet/v2/pkg/config"
+
+// Configure configures individual resources by adding custom ResourceConfigurators.
+func Configure(p *config.Provider) {
+	p.AddResourceConfigurator("openstack_lb_quota_v2", func(r *config.Resource) {
+		r.References["project_id"] = config.Reference{
+			TerraformName: "openstack_identity_project_v3",
+		}
+	})
+}