+++ config/external_name.go
@@ -1,6 +1,6 @@
 package config
 
-import "github.com/crossplane/upjet/pkg/config"
+import "github.com/crossplane/upjet/v2/pkg/config"
 
 // ExternalNameConfigs contains all external name configurations for this
 // provider.